+++ src/client/components/home/Home.js
@@ -3,7 +3,7 @@ import NavbarContainer from "../navbar/NavbarContainer";
 import { Link } from "react-router-dom";
 import SubNav from "./SubNav";
 import PropTypes from "prop-types";
-import ModalFirstAccess from "./ModalFirstAccess";
+import ModalFirstAccess from "../User/ModalFirstAccess";
 
 const Home = ({ user }) => {
   let firstAccess = typeof user !== "undefined" ? user.firstAccess : false;
